fix: resolve Phase B warnings — weights spec and key unification
Warning 1 — aggregator.py: correct fallback hardcoded weights
  Old: calibration=0.25, semantic_uncertainty=0.30, cross_check=0.45
  New: calibration=0.20, semantic_uncertainty=0.50, cross_check=0.30
  These match the spec and are only used if config.yaml is missing
  the detection.weights block (never triggered in practice).

Warning 2 — aggregator.py: unify weights_used key name
  Rename 'uncertainty' → 'semantic_uncertainty' in HallucinationResult.
  This makes the key consistent across: aggregator, explainer, API,
  UI components, and tests. Remove the explicit rename workaround
  in run_full_pipeline() that papered over the inconsistency.

Warning 3 — model_loader.py: fix torch_dtype deprecation
  transformers 5.x renamed torch_dtype → dtype in from_pretrained().
  Use dtype= to suppress the deprecation warning.

+++ models/model_loader.py
@@ -56,7 +56,7 @@ def get_open_model():
             model_name,
             quantization_config=bnb_config,
             device_map="auto" if use_4bit else None,
-            torch_dtype=dtype,
+            dtype=dtype,               # torch_dtype renamed to dtype in transformers 5.x
             trust_remote_code=True
         )
         if not use_4bit:
